Remove the unnecessary hook and moved everything to the RemoteComponent file
