+++ .gitignore
@@ -1,4 +1,4 @@
-/node_modules
+node_modules
 /.pnp
 .pnp.js
 
@@ -11,6 +11,7 @@
 
 # production
 /build
+/dist
 
 # misc
 .DS_Store
